DPR_MedFusionNet/requirements.txt
# Local inference requirements for the notebook-v4 MedFusionNet model
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0
